refactor(models): fold starting-equipment into classes.startingEquipment
Equipment resolution is tightly coupled to class data — every
StartingEquipmentItem lives on a ClassDetails. It belongs as a
projection on the classes model, not a standalone helper file.

- Move the toEquipment / resolveStartingEquipment logic into classes.ts.
- Expose classes.startingEquipment({id}) returning Equipment[].
- Delete src/models/common/gear/starting-equipment.ts + its test.
- Migrate character-creation.tsx and step-equipment.tsx consumers.
- Port the test cases to classes.test.ts as classes.startingEquipment()
  describes, plus a new assertion that each resolved armor piece
  carries its armorId (from the step-4 change).
