+++ src/models/class/classes.test.ts
@@ -72,3 +72,50 @@ describe("classes.groupBy()", () => {
     ).toThrow(/Unknown group axis/);
   });
 });
+
+describe("classes.startingEquipment()", () => {
+  it("returns equipment for fighter", () => {
+    const equipment = classes.startingEquipment({ id: "fighter" });
+    expect(equipment.length).toBeGreaterThan(0);
+  });
+
+  it("resolves weapons with damage", () => {
+    const equipment = classes.startingEquipment({ id: "fighter" });
+    const weapons = equipment.filter((e) => e.type === "weapon");
+    expect(weapons.length).toBeGreaterThan(0);
+    for (const weapon of weapons) {
+      expect(weapon.damage).toBeDefined();
+    }
+  });
+
+  it("resolves armor with ac", () => {
+    const equipment = classes.startingEquipment({ id: "fighter" });
+    const armor = equipment.filter((e) => e.type === "armor");
+    expect(armor.length).toBeGreaterThan(0);
+    for (const piece of armor) {
+      expect(piece.ac).toBeDefined();
+      expect(piece.armorId).toBeDefined();
+    }
+  });
+
+  it("resolves shields as type 'shield' not 'armor'", () => {
+    const equipment = classes.startingEquipment({ id: "cleric" });
+    const shields = equipment.filter((e) => e.name === "Shield");
+    expect(shields).toHaveLength(1);
+    expect(shields[0].type).toBe("shield");
+  });
+
+  it("resolves gold as type 'money'", () => {
+    const equipment = classes.startingEquipment({ id: "fighter" });
+    const money = equipment.filter((e) => e.type === "money");
+    expect(money).toHaveLength(1);
+    expect(money[0].name).toBe("Gold");
+    expect(money[0].quantity).toBeGreaterThan(0);
+  });
+
+  it("returns equipment for all classes without throwing", () => {
+    for (const cls of classes.list()) {
+      expect(() => classes.startingEquipment({ id: cls.id })).not.toThrow();
+    }
+  });
+});
